+++ NEWS.md
@@ -1,3 +1,5 @@
+# ggstats (development version)
+
 # ggstats 0.1.1
 
 * Examples relying on Internet resources have been removed (#11)